Added wildlife attractor

+++ src/main/java/com/bartz24/skyresources/technology/item/ItemRockGrinder.java
@@ -36,7 +36,7 @@ public class ItemRockGrinder extends ItemPickaxe
 		super(material);
 		toolMaterial = material;
 		this.setMaxDamage((int) (material.getMaxUses() * ConfigOptions.toolSettings.rockGrinderBaseDurability));
-		this.damageVsEntity = ConfigOptions.toolSettings.rockGrinderBaseDamage + material.getDamageVsEntity();
+		this.damageVsEntity = (float) (ConfigOptions.toolSettings.rockGrinderBaseDamage + material.getDamageVsEntity());
 		this.setUnlocalizedName(References.ModID + "." + unlocalizedName);
 		setRegistryName(registryName);
 		this.setMaxStackSize(1);
